chapter-10/02-time-detectors/01-lag-detector-delta-lake-apache-spark-grafana/requirements.txt
delta-spark==3.0.0
pyspark==3.5.0
prometheus-client==0.20.0